requirements.txt
torch~=2.1.2
numpy==1.26.3
torchvision==0.16.2
transformers==4.36.2
requests==2.31.0
pillow==10.2.0
monai==1.3.0
matplotlib==3.8.0
pandas==2.1.4
scikit-learn==1.3.0